requirements/common.txt
coverage==3.7.1
dj-static==0.0.6
Django==1.7.4
django-angular==0.7.10
django-debug-toolbar==1.2.2
django-dynamic-fixture==1.8.1
django-filter==0.9.2
django-postgrespool==0.3.0
django-redis-cache==0.13.0
django-social-auth==0.7.28
django-toolbelt==0.0.1
djangorestframework==3.0.5
gunicorn==19.2.1
httplib2==0.9
Markdown==2.6
oauth2==1.5.211
python-openid==2.2.5
redis==2.10.3
selenium==2.44.0
six==1.9.0
SQLAlchemy==0.9.8
sqlparse==0.1.14
static3==0.5.1
python-dateutil==2.4.0
Jinja2==2.7.3
Unidecode==0.4.17
django-nvd3==0.7.7
markupsafe==0.23
python-nvd3==0.13.5
python-slugify==0.0.9
django-bower==5.0.2